requirements.txt
python-binance
python-dotenv
pandas
streamlit
streamlit-autorefresh
matplotlib
